requirements.txt
hydra-core==1.2.0
PySoundFile==0.9.0.post1
python-dotenv==0.21.0
torch==1.11.0
torchaudio==0.11.0
torch-summary==1.4.5
torchvision==0.12.0
tqdm==4.64.1
wandb==0.12.21